Use only Tini for init and signal passing
With the addition of `confd` outside the prep_term block, the prep_term and wait_term are not existing the container when Kamailio crashes as the signal isn't being passed properly.

However, prep_term and wait_term are no longer necessary since tini is being used for the container init. Tini iitself should be replaced with the tini built into docker-ce now with the `--init` flag.